Fix race conditions during exiting.
A side effect of this change is memory usage increase because it moves
ended channel to a vector instead of dispose them immediately.

This change also add an reference to pc_factory_ when it is created. It
causes memory leak but fixes hang issue when application exits.
AtExitManager might be needed to eventually fix this issue.